requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dateutil==2.8.2
python-dotenv==1.0.1
